+++ src/PostCell.swift
@@ -54,8 +54,13 @@ class PostCell: UITableViewCell, AvatarImageViewDelegate {
 
         if let topBin = topBin {
             emptyOut(topBin)
+
             headerView.map { topBin.addArrangedSubview($0) }
-            avatarToTopBin?.constant = (headerView != nil) ? 8 : 0
+
+            // Place any title last, so it's right by the post content.
+            self.buildPostTitleLabel(showing: post.title).map { topBin.addArrangedSubview($0) }
+
+            avatarToTopBin?.constant = topBin.arrangedSubviews.isEmpty ? 0 : 8
         }
 
         stackUpAdditionalInfo()
@@ -86,6 +91,20 @@ class PostCell: UITableViewCell, AvatarImageViewDelegate {
         formatter.doesRelativeDateFormatting = true
         return formatter
     }()
+
+
+    // MARK: - Injects a title label above everything
+    func buildPostTitleLabel(showing title: String) -> UILabel? {
+        let title = title.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !title.isEmpty else { return nil }
+
+        let label = UILabel()
+        label.font = UIFont.preferredFont(forTextStyle: .title1, compatibleWith: self.traitCollection)
+        label.numberOfLines = 0
+        enableAutoContentSizeUpdates(for: label)
+        label.text = title
+        return label
+    }
 
 
     // MARK: - Alters background color to reflect mention status