+++ CHANGELOG.md
@@ -8,6 +8,16 @@ Fixed:
 
 - Crash when attempting to show a 10Centuries error message.
 
+Changed:
+
+- Supports 10Cv5 rather than 10Cv4.
+
+New:
+
+- Displays any non-empty post title above the post.
+- Displays any post geo-tag info in post attributes.
+
+
 ### 1.0 (22) - 2019-03-29
 Removed: